Backend/requirements.txt
django

langchain
langchain-core
langchain_community
langchain-google-genai
langchain-chroma
youtube-transcript-api

pypdf

python-dotenv